pokernow-analyzer-web/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
supabase==2.0.3
python-dotenv==1.0.0
